fix bugprone-* * bugprone-throw-keyword-missing * bugprone-unhandled-self-assignment * bugprone-suspicious-string-compare * bugprone-reserved-identifier * bugprone-narrowing-conversions * bugprone-macro-parentheses * bugprone-implicit-widening-of-multiplication-result * bugprone-exception-escape * bugprone-copy-constructor-init

+++ src/Base/RotationPyImp.cpp
@@ -674,7 +674,7 @@ PyObject* RotationPy::number_power_handler(PyObject* self, PyObject* other, PyOb
     double rfAngle {};
 
     a.getRawValue(axis, rfAngle);
-    rfAngle *= b;
+    rfAngle *= double(b);
     a.setValue(axis, rfAngle);
 
     return new RotationPy(a);
